% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissqlite-package.R
\docType{package}
\name{poissqlite-package}
\alias{poissqlite}
\alias{poissqlite-package}
\title{poissqlite: Work with SQLite Databases}
\description{
Facilitates working with SQLite databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/poissqlite}
  \item Report bugs at \url{https://github.com/poissonconsulting/poissqlite/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

Authors:
\itemize{
  \item Seb Dalgarno
}

}
\keyword{internal}
