% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{poissqlite}
\alias{poissqlite}
\alias{package-poissqlite}
\alias{poissqlite-package}
\alias{poissqlite-package}
\title{poissqlite}
\description{
Facilitates working with SQLite databases.
}
\seealso{
\url{https://github.com/poissonconsulting/poissqlite}
}
