% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob}
\alias{ps_blob}
\title{BLOB}
\usage{
ps_blob(dir = ".", pattern = "[.]pdf$", n = 10000L, recursive = TRUE)
}
\arguments{
\item{dir}{A string of the directory name.}

\item{pattern}{A string of the pattern to use when searching for files.}

\item{n}{An integer of the (maximal) number of records to be read.}

\item{recursive}{A flag indicating whether to recurse into subdirectories.}
}
\description{
Converts files in the directory into a named list of BLOBs.
}
\details{
For importing into SQLite database it is easiest to convert to a tibble
using \code{\link{ps_blob_to_tibble}}.

The file extension is stored inside each blob so there is no need to save the filenames.
}
\seealso{
\code{\link{ps_deblob}}, \code{\link{ps_blob_to_tibble}} and \code{\link{readBin}}
}
