% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob_files}
\alias{ps_blob_files}
\title{Blob Files}
\usage{
ps_blob_files(dir = ".", pattern = "[.]pdf$", recursive = FALSE)
}
\arguments{
\item{dir}{A string of the directory name.}

\item{pattern}{A string of the pattern to use when searching for files.}

\item{recursive}{A flag indicating whether to recurse into subdirectories.}
}
\description{
Converts files in the directory into a tibble with a column File of file names
and a column BLOB of blob objects.
}
