% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob_object}
\alias{ps_blob_object}
\title{Blob Object}
\usage{
ps_blob_object(object, name = substitute(object))
}
\arguments{
\item{object}{An R object.}

\item{name}{A string of the name for the blob.}
}
\description{
Converts an R object into a named blob scalar.
}
