% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob_to_tibble}
\alias{ps_blob_to_tibble}
\title{BLOB to tibble}
\usage{
ps_blob_to_tibble(x)
}
\arguments{
\item{x}{A list of blobs}
}
\value{
A tibble with columns File and BLOB.
}
\description{
BLOB to tibble
}
