% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{ps_column_info}
\alias{ps_column_info}
\title{SQL column info}
\usage{
ps_column_info(table_name, conn = getOption("ps.conn"))
}
\arguments{
\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\value{
A data frame with the name of the column and the type.
}
\description{
SQL column info
}
