% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ps_connect_sqlite}
\alias{ps_connect_sqlite}
\title{Opens a connection to an sqlite database.}
\usage{
ps_connect_sqlite(file = "database", dir = ".", new = NA,
  foreign_keys = TRUE, ask = getOption("poissqlite.ask", TRUE))
}
\arguments{
\item{file}{A string of the name of the file (without the file extension .sqlite).}

\item{dir}{A string of the directory.}

\item{new}{A flag indicating whether to enforce connection to an existing database (FALSE) or connection to a new database (TRUE).}

\item{foreign_keys}{A flag indicating whether to switch foreign keys on.}

\item{ask}{A flag indicating whether to ask before creating a new directory.}
}
\description{
The connection can be queried and closed using the functions
in the \code{DBI} package.
}
