% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{ps_create_sql_script}
\alias{ps_create_sql_script}
\title{Create sql database}
\usage{
ps_create_sql_script(x, db_name = "", load = "prepare",
  path = "create-database.R")
}
\arguments{
\item{x}{A named list of data.frames.}

\item{db_name}{A character string of the name of the database.}

\item{load}{A character string indicating subfolder to load data from.}

\item{path}{A character string indicating path and file name of the script.}
}
\value{
A R script to create a sql database.
}
\description{
Extracts information from a list of data.frames to create code to write an sql database.
Tables are added to the script in the same order as the named list
}
