% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_deblob}
\alias{ps_deblob}
\title{DeBLOB}
\usage{
ps_deblob(x, dir = ".")
}
\arguments{
\item{x}{A list of BLOBs created by \code{\link{ps_blob}}.}

\item{dir}{A string of the directory to save the files to.}
}
\value{
An invisible vector of the names of the files saved to dir.
}
\description{
Converts a possible named list of BLOBs into files in the directory.
}
\details{
If x is unnamed the files are assigned the names
file1, file2 etc according to their order in x.
}
\seealso{
\code{\link{ps_deblob}}
}
