% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_deblob_file}
\alias{ps_deblob_file}
\title{Deblob to File}
\usage{
ps_deblob_file(blob, file = "blob", dir = ".",
  ask = getOption("poissqlite.ask", TRUE))
}
\arguments{
\item{blob}{A raw object.}

\item{file}{A string of the file (the original file extension is added automatically)}

\item{dir}{A string of the directory.}

\item{ask}{A flag indicating whether to ask before creating the directory or replacing a file.}
}
\description{
Converts a raw object back to its original file format.
}
