% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_deblob_object}
\alias{ps_deblob_object}
\title{Deblob Object}
\usage{
ps_deblob_object(blob)
}
\arguments{
\item{blob}{A raw object.}
}
\description{
Converts a raw object into its original file format
 and if it is an .rds file reads it it as an R object. Otherwise it throws an error.
}
\examples{
mat <- matrix(1:9, nrow = 3)
blob <- ps_blob_object(mat)
ps_deblob_object(blob)
}
