% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{ps_delete_data}
\alias{ps_delete_data}
\title{Delete Data}
\usage{
ps_delete_data(table_name, conn = getOption("ps.conn"))
}
\arguments{
\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\description{
Deletes data in an existing table.
}
