% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{ps_df_to_sql}
\alias{ps_df_to_sql}
\title{Data.frame to sql}
\usage{
ps_df_to_sql(data, data_name = deparse(substitute(data)),
  table_name = tools::toTitleCase(data_name))
}
\arguments{
\item{data}{A data.frame.}

\item{data_name}{A character string of the name of the data.frame.}

\item{table_name}{A character string of the name of the sql table.}
}
\value{
A character string.
}
\description{
Draws information from a data.frame to provide code to write an sql table.
}
