% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ps_disconnect_sqlite}
\alias{ps_disconnect_sqlite}
\title{Close a connection to an sqlite database.}
\usage{
ps_disconnect_sqlite(conn)
}
\arguments{
\item{conn}{The connection to close.}
}
\description{
A wrapper on DBI::dbDisconnect.
}
