% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{ps_find_key}
\alias{ps_find_key}
\title{Find primary key}
\usage{
ps_find_key(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A vector of column names
}
\description{
Uses a simple algorithm to search for a likely primary key.
}
