% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ps_interpret_data}
\alias{ps_interpret_data}
\title{Interpret Data}
\usage{
ps_interpret_data(x, table_name, conn = getOption("ps.conn"))
}
\arguments{
\item{x}{A data frame of columns to interpret.}

\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\value{
The original data frame with units and column comments.
}
\description{
Takes a data frame and sets the units and columns for columns which appear
in the metadata for table table_name.
}
