% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ps_load_tables}
\alias{ps_load_tables}
\title{Load Tables}
\usage{
ps_load_tables(conn = getOption("ps.conn"), rename = identity,
  envir = parent.frame())
}
\arguments{
\item{conn}{An SQLiteConnection object.}

\item{rename}{A function to alter the SQLite database table names.}

\item{envir}{The environment to assign the tables to.}
}
\value{
An invisible vector of table names.
}
\description{
Assigns tables in an SQLite database to environment
as tibble or sf objects (if geometry column).
}
