% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ps_read_table}
\alias{ps_read_table}
\title{Read Table}
\usage{
ps_read_table(table_name, conn = getOption("ps.conn"))
}
\arguments{
\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\description{
Returns a table in an SQLite database as a tibble or sf object.
}
