% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ps_read_tables}
\alias{ps_read_tables}
\title{Read Tables}
\usage{
ps_read_tables(conn, rename = identity)
}
\arguments{
\item{conn}{An SQLiteConnection object.}

\item{rename}{A function to alter the SQLite database table names.}
}
\description{
Assigns tables in an SQLite database to global environment
as tibble or sf objects.
}
