% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-columns.R
\name{ps_strip_columns}
\alias{ps_strip_columns}
\title{Strip Columns}
\usage{
ps_strip_columns(columns, envir = parent.frame())
}
\arguments{
\item{columns}{A character vector of the columns to remove (if present).}

\item{envir}{The environment.}
}
\value{
An invisible vector of the modified data frames.
}
\description{
Removes columns from all data frames in envir.
}
