% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definition.R
\name{ps_table_definition}
\alias{ps_table_definition}
\title{Table Definition}
\usage{
ps_table_definition(table_name, conn = getOption("ps.conn"))
}
\arguments{
\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\value{
A string of the SQL definition of the table.
}
\description{
Table Definition
}
