% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ps_write_table}
\alias{ps_write_table}
\title{Write Table}
\usage{
ps_write_table(x, table_name, conn)
}
\arguments{
\item{x}{The data frame.}

\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}
}
\description{
Appends a data frame to an existing table in an SQLite database without row names.
}
