% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-tables-csv.R
\name{ps_write_tables_csvs}
\alias{ps_write_tables_csvs}
\title{Write Tables}
\usage{
ps_write_tables_csvs(conn = getOption("ps.conn"), dir = ".",
  rename = identity)
}
\arguments{
\item{conn}{An SQLiteConnection object.}

\item{dir}{A string of the directory name.}

\item{rename}{A function to alter the SQLite database table names.}
}
\value{
An invisible character vector of the data names.
}
\description{
Writes all the tables in the sqlite database conn to directory dir
using \code{write_csv}.
}
