% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_comment}
\alias{set_comment}
\title{Set Comment}
\usage{
set_comment(x, comment)
}
\arguments{
\item{x}{The object}

\item{comment}{A string of the comment}
}
\value{
The modified object
}
\description{
Set Comment
}
\examples{
x <- 1:2
x <- set_comment(x, "A vector")
comment(x)
}
