% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_blob}
\alias{ps_blob}
\title{BLOB}
\usage{
ps_blob(dir = ".", pattern = "[.]pdf$", n = 10000L, recursive = TRUE)
}
\arguments{
\item{dir}{A string of the directory name.}

\item{pattern}{A string of the pattern to use when searching for files.}

\item{n}{An integer of the (maximal) number of records to be read.}

\item{recursive}{A flag indicating whether to recurse into subdirectories.}
}
\value{
A named list of the BLOBs.
}
\description{
Converts files in the directory into a list of BLOBs for storage in an SQLite database.
}
\seealso{
\code{\link{ps_deblob}}
}
