% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob.R
\name{ps_deblob_object}
\alias{ps_deblob_object}
\title{Deblob Object}
\usage{
ps_deblob_object(blob)
}
\arguments{
\item{blob}{A blob.}
}
\description{
Converts a blob into an R object. Throws an error if the object was not created using
ps_blob_object.
}
\examples{
mat <- matrix(1:9, nrow = 3)
blob <- ps_blob_object(mat)
ps_deblob_object(blob)
}
