% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ps_write_table}
\alias{ps_write_table}
\title{Write Table}
\usage{
ps_write_table(x, table_name, conn = getOption("ps.conn"), delete = FALSE,
  overwrite_units = FALSE, overwrite_descriptions = FALSE,
  rename = identity)
}
\arguments{
\item{x}{The data frame.}

\item{table_name}{A string of the name of the table.}

\item{conn}{An SQLiteConnection object.}

\item{delete}{A flag indicating whether to delete the existing data before appending the data.}

\item{overwrite_units}{A flag indicating whether to overwrite existing units.}

\item{overwrite_descriptions}{A flag indicating whether to overwrite existing descriptions.}

\item{rename}{A function to rename column names in x.}
}
\description{
Appends a data frame to an existing table in an SQLite database
without row names.
More importantly it saves the time zone for POSIXct columns
in the metadata table and converts the column to a character vector
of format `YYYY-mm-dd HH:MM:SS`.
It also saves the projection and converts the geometry column to a character
vector for sfc columns. And saves any column comments to the DataDescription
entry in the metadata table.
}
\details{
\code{\link{ps_read_table}} looks up the time zone and projection and
converts the column back to a POSIXct vector with the original timezone
or sfc column with projection.

The function is also agnostic as to the order of the columns.
}
